+++ python/cugraph/cugraph/experimental/link_analysis/hits.py
@@ -1,4 +1,5 @@
-# Copyright (c) 2019-2022, NVIDIA CORPORATION.
+# Copyright (c) 2022, NVIDIA CORPORATION.
+#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -10,14 +11,22 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+#
 
-from cugraph.link_analysis import hits_wrapper
 from cugraph.utilities import (ensure_cugraph_obj_for_nx,
                                df_score_to_dictionary,
                                )
+from pylibcugraph import (ResourceHandle,
+                          GraphProperties,
+                          SGGraph,
+                          hits as pylibcugraph_hits
+                          )
+import cudf
 
 
-def hits(G, max_iter=100, tol=1.0e-5, nstart=None, normalized=True):
+def EXPERIMENTAL__hits(
+    G, max_iter=100, tol=1.0e-5, nstart=None, normalized=True
+):
     """
     Compute HITS hubs and authorities values for each vertex
 
@@ -25,35 +34,33 @@ def hits(G, max_iter=100, tol=1.0e-5, nstart=None, normalized=True):
     estimates the node value based on the incoming links.  Hubs estimates
     the node value based on outgoing links.
 
-    The cuGraph implementation of HITS is a wrapper around the gunrock
-    implementation of HITS.
-
-    Note that the gunrock implementation uses a 2-norm, while networkx
-    uses a 1-norm.  The raw scores will be different, but the rank ordering
-    should be comparable with networkx.
+    Both cuGraph and networkx implementation use a 1-norm.
 
     Parameters
     ----------
-    graph : cugraph.Graph
+    G : cugraph.Graph
         cuGraph graph descriptor, should contain the connectivity information
         as an edge list (edge weights are not used for this algorithm).
         The adjacency list will be computed if not already present.
 
     max_iter : int, optional (default=100)
         The maximum number of iterations before an answer is returned.
-        The gunrock implementation does not currently support tolerance,
-        so this will in fact be the number of iterations the HITS algorithm
-        executes.
 
     tol : float, optional (default=1.0e-5)
         Set the tolerance the approximation, this parameter should be a small
-        magnitude value.  This parameter is not currently supported.
+        magnitude value.
 
     nstart : cudf.Dataframe, optional (default=None)
-        Not currently supported
+        The initial hubs guess vertices along with their initial hubs guess
+        value
+
+        nstart['vertex'] : cudf.Series
+            Initial hubs guess vertices
+        nstart['values'] : cudf.Series
+            Initial hubs guess values
 
     normalized : bool, optional (default=True)
-        Not currently supported, always used as True
+        A flag to normalize the results
 
     Returns
     -------
@@ -82,15 +89,44 @@ def hits(G, max_iter=100, tol=1.0e-5, nstart=None, normalized=True):
 
     G, isNx = ensure_cugraph_obj_for_nx(G)
 
-    df = hits_wrapper.hits(G, max_iter, tol)
+    srcs = G.edgelist.edgelist_df['src']
+    dsts = G.edgelist.edgelist_df['dst']
+    # edge weights are not used for this algorithm
+    weights = G.edgelist.edgelist_df['src'] * 0.0
 
-    if G.renumbered:
-        df = G.unrenumber(df, "vertex")
+    resource_handle = ResourceHandle()
+    graph_props = GraphProperties(is_multigraph=G.is_multigraph())
+    store_transposed = False
+    renumber = False
+    do_expensive_check = False
+    init_hubs_guess_vertices = None
+    init_hubs_guess_values = None
+
+    if nstart is not None:
+        init_hubs_guess_vertices = nstart['vertex']
+        init_hubs_guess_values = nstart['values']
+
+    sg = SGGraph(resource_handle, graph_props, srcs, dsts, weights,
+                 store_transposed, renumber, do_expensive_check)
+
+    vertices, hubs, authorities = pylibcugraph_hits(resource_handle, sg, tol,
+                                                    max_iter,
+                                                    init_hubs_guess_vertices,
+                                                    init_hubs_guess_values,
+                                                    normalized,
+                                                    do_expensive_check)
+    results = cudf.DataFrame()
+    results["vertex"] = cudf.Series(vertices)
+    results["hubs"] = cudf.Series(hubs)
+    results["authorities"] = cudf.Series(authorities)
 
     if isNx is True:
-        d1 = df_score_to_dictionary(df[["vertex", "hubs"]], "hubs")
-        d2 = df_score_to_dictionary(df[["vertex", "authorities"]],
+        d1 = df_score_to_dictionary(results[["vertex", "hubs"]], "hubs")
+        d2 = df_score_to_dictionary(results[["vertex", "authorities"]],
                                     "authorities")
-        df = (d1, d2)
+        results = (d1, d2)
 
-    return df
+    if G.renumbered:
+        results = G.unrenumber(results, "vertex")
+
+    return results
